// src/api/generated/model/fuelStationRecommendationResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MiniCRM API
 * API Documentation for MiniCRM Application
 * OpenAPI spec version: 1.0.0
 */
import type { CalculatedCostDto } from './calculatedCostDto';

export interface FuelStationRecommendationResponse {
  id?: number;
  name?: string;
  address?: string;
  distanceKm?: number;
  pricePerLiter?: number;
  calculatedCost?: CalculatedCostDto;
  googleMapsUrl?: string;
  isOpen?: boolean;
}